Merge r1773346 from trunk:
Drop C-L header and message-body from HTTP 204 responses.

The C-L header can be set in a fcgi/cgi backend or in other
filters like ap_content_length_filter (with the value of 0),
meanwhile the message-body can be returned incorrectly
by any backend. The idea is to remove unnecessary bytes
from a HTTP 204 response.

PR 51350


Note: merged to avoid manual conflicts, became a depedendency of the HTTP 
strict in trunk.



git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x-merge-http-strict@1775668 13f79535-47bb-0310-9956-ffa450edef68